+++ include/aligator/solvers/fddp/solver-fddp.hxx
@@ -23,8 +23,8 @@ SolverFDDPTpl<Scalar>::SolverFDDPTpl(const Scalar tol, VerboseLevel verbose,
 template <typename Scalar>
 void SolverFDDPTpl<Scalar>::setup(const Problem &problem) {
   problem.checkIntegrity();
-  results_ = Results(problem);
-  workspace_ = Workspace(problem);
+  new (&results_) Results(problem);
+  new (&workspace_) Workspace(problem);
   // check if there are any constraints other than dynamics and throw a warning
   std::vector<std::size_t> idx_where_constraints;
   for (std::size_t i = 0; i < problem.numSteps(); i++) {